q-note/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
websockets==13.1
httpx==0.27.2
openai==1.54.4
pydantic==2.9.2
python-multipart==0.0.12
trafilatura==1.12.2
pdfplumber==0.11.4
python-docx==1.1.2
resemblyzer==0.1.4
scikit-learn==1.5.2
# torch 는 CPU-only: pip install torch==2.5.1 --index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
numpy>=1.24
